+++ packages/core/CHANGELOG.md
@@ -3,6 +3,64 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.33.0](https://github.com/medly/medly-components/compare/@medly-components/core@1.32.3...@medly-components/core@1.33.0) (2020-10-30)
+
+
+### Bug Fixes
+
+* add fullwidth to Textfield component in Datepicker ([5ee41df](https://github.com/medly/medly-components/commit/5ee41df34ff243e6698b22f119e4c483001baf4d))
+* padding issue in Textfield component ([43b7e25](https://github.com/medly/medly-components/commit/43b7e255bfb58cbf6002821db40a02b6f4841614))
+* **core:** add missing dependency in datepicker text change callback definition ([b1fcee3](https://github.com/medly/medly-components/commit/b1fcee37777fae37c941fc31c365fb31036c783a))
+* improve test coverage, attempt mobile bugfix ([7aa927b](https://github.com/medly/medly-components/commit/7aa927bb4554a9dd41d0e89be80c9aabaa445a57))
+* update snapshots ([21b7060](https://github.com/medly/medly-components/commit/21b7060ef1709cbc473cc6ffb372149f4ae7049a))
+
+
+### Features
+
+* added formatter support for custom component in table row ([660720f](https://github.com/medly/medly-components/commit/660720f3b7646bab6e33196563e32952a2a719f1))
+* adding fixes for existing issues, and addressing pr comments ([b5ce890](https://github.com/medly/medly-components/commit/b5ce89048edb4be4e304190ee07ccdc707847bc2))
+* adding mobile functionality for the datepicker component ([338bf77](https://github.com/medly/medly-components/commit/338bf77f82c74becdd5b9fd4420ad051c0f309c7))
+* adding more requested PR changes ([42042fa](https://github.com/medly/medly-components/commit/42042fa56f33c374c281b5d1b6c37c157197f2aa))
+* adding value functionality ([70358fa](https://github.com/medly/medly-components/commit/70358fa7bce8397bc0a18010be0657c1c95f9a8b))
+
+
+
+
+
+## [1.32.3](https://github.com/medly/medly-components/compare/@medly-components/core@1.32.2...@medly-components/core@1.32.3) (2020-10-29)
+
+
+### Bug Fixes
+
+* width and height issue in Modal component ([465d4b6](https://github.com/medly/medly-components/commit/465d4b6104a1e879bf739046d0d64b423ed7e99a))
+
+
+
+
+
+## [1.32.2](https://github.com/medly/medly-components/compare/@medly-components/core@1.32.1...@medly-components/core@1.32.2) (2020-10-29)
+
+
+### Bug Fixes
+
+* prop type for groupBy in Table component ([85c9bae](https://github.com/medly/medly-components/commit/85c9bae7e41b09a5fbba887710c9bbc12444a475))
+
+
+
+
+
+## [1.32.1](https://github.com/medly/medly-components/compare/@medly-components/core@1.32.0...@medly-components/core@1.32.1) (2020-10-28)
+
+
+### Bug Fixes
+
+* add autofill css for mozilla ([9300b1f](https://github.com/medly/medly-components/commit/9300b1fff4eae432be3c11fb58b11783c45c28ce))
+* update value on change of the options ([56f1db0](https://github.com/medly/medly-components/commit/56f1db0817ad3cd259ddc59b67a8e724f46ac808))
+
+
+
+
+
 # [1.32.0](https://github.com/medly/medly-components/compare/@medly-components/core@1.31.2...@medly-components/core@1.32.0) (2020-10-28)